+++ Semana_2/node/src/server.ts
@@ -1,38 +1,42 @@
-import express, {Request,Response,NextFunction} from "express";
-import  "express-async-errors"; // Para poder tratar errors async, deve estar abaixo da importacao do express
+import express, { Request, Response, NextFunction } from "express";
+import "express-async-errors"; // Para poder tratar errors async, deve estar abaixo da importacao do express
 import routes from "./routes";
-import 'reflect-metadata'
-import './database'
-import uploadConfig from './config/upload'
-import AppError from './errors/AppError'
+import Cors from "cors"; // Serve apenas para apps com interface web
+import "reflect-metadata";
+import "./database";
+import uploadConfig from "./config/upload";
+import AppError from "./errors/AppError";
 
 const app = express();
-app.use('/files',express.static(uploadConfig.directory))
-app.use(routes)
+
+app.use(Cors({
+  origin: 'http://localhost:3000'// Para ambiente de dev 
+}));// Evita que sites nao confiaveis a aplicacao acessem a api.
+app.use("/files", express.static(uploadConfig.directory));
+app.use(routes);
 
 //app.use(routes)
 app.get("/", (request, response) => {
   return response.json({ message: "Hello World" });
 });
 
-app.use((err:Error,request:Request,response:Response,next:NextFunction) =>{
-  if(err instanceof AppError){
-    return response.status(err.statusCode).json({
-      status:"error",
-      message: err.message
-    })
+app.use(
+  (err: Error, request: Request, response: Response, next: NextFunction) => {
+    if (err instanceof AppError) {
+      return response.status(err.statusCode).json({
+        status: "error",
+        message: err.message,
+      });
+    }
+
+    console.log(err);
+
+    return response.status(500).json({
+      status: "error",
+      message: "Internal server error",
+    });
   }
-
-  console.log(err)
-
-  return response.status(500).json({
-    status: "error",
-    message: "Internal server error",
-  })
-}) //Middlewares para tratativa de erros no express devem receber 4 parametros e sempre ficar apos a utilizacao das demais rotas
-
-
-
+); //Middlewares para tratativa de erros no express devem receber 4 parametros e sempre ficar apos a utilizacao das demais rotas
 
 app.listen(3333, () => {
   console.log("Running 🚀");